Added UI with Slint
I was able to get Slint working where it could display the information needed for the app.

There is a weird bug, wher when you run the program the console remains open.
The other complaint about slint is that the UI is kind of ugly / hard to work with.
A feature id like to add seperate the UI into its own UI folder etc, but i would need to do more practicing with slint ouside of the app.
The final complaint is that I would like to add the abbility to update the values in the UI